fixed logic errors
the Tkinter library stopped being included in the start of the file as well as some functions to change the absolute and local location/orientations of the camera. this allows for the camera to be properly moved instead of static like previously before. these functions also now include math.radians() so degrees can be used to rotate the player.