Update Main.py
Implemented a version of the painters algorithm in order to create the proper rendering order for polygons. This was done by determining the furthest z point from the camera and sorting the polygons via that. NOTE: could arise some ordering issues with really large polygons that should be rendered on top but have points far back.